requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.4
django-cors-headers==3.5.0
djangorestframework==3.12.2
drf-yasg==1.20.0
googlemaps==4.4.2
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
packaging==20.7
pandas==1.1.4
psycopg2==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
tqdm==4.54.0
uritemplate==3.0.1
urllib3==1.26.2
